[Refactor] Update action packages and improve action execution methods
