+++ src/main/java/com/vanixmc/events/action/core_actions/ClearVariableAction.java
@@ -1,4 +1,4 @@
-package com.vanixmc.events.action.impl.clear_variable_action;
+package com.vanixmc.events.action.core_actions;
 
 import com.vanixmc.events.action.domain.AbstractAction;
 import com.vanixmc.events.context.Context;
