+++ src/main/java/com/vanixmc/events/action/core_actions/command_action/CommandSender.java
@@ -1,0 +1,6 @@
+package com.vanixmc.events.action.core_actions.command_action;
+
+public enum CommandSender {
+    PLAYER,
+    CONSOLE
+}
