--- src/main/java/com/vanixmc/events/action/impl/command_action/CommandSender.java
@@ -1,6 +1,0 @@
-package com.vanixmc.events.action.impl.command_action;
-
-public enum CommandSender {
-    PLAYER,
-    CONSOLE
-}
